requirements.txt
Django==1.6.2
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-crispy-forms==1.4.0
django-floppyforms==1.1
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
selenium==2.41.0
static==1.0.2
wsgiref==0.1.2
